vercel_requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
gunicorn==21.2.0
psycopg2-binary==2.9.9
Pillow==10.1.0
imagehash==4.3.1
requests==2.31.0
authlib==1.2.1
oauthlib==3.2.2
email-validator==2.1.0
openai==1.3.8
replicate==0.22.0
reportlab==4.0.7
qrcode==7.4.2
Werkzeug==3.0.1
SQLAlchemy==2.0.23